primera implementacion de ciberseguridad

+++ src/main/java/com/cyberpath/springboot/dto/ejercicio/IntentoEjercicioDto.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -13,7 +12,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class IntentoEjercicioDto {
     private Integer id;
-    private BigDecimal puntaje;
+    private Double puntaje;
     private Integer duracionSeg;
     private String fecha;
     private String estado;
